RF: move and make fixtures available outside of the dandi/tests
